Remove dynamic lights in the lobby
Causes visual glitches and blackouts when changing modes or staying on the lobby for a while. If we want custon light we should probably reverse engineer tweaklights and adding these in for the lobby as the lobby seems to deal with them differently. Removing this will ensure the lobby looking properly for all users.