Comment out MOTD duplicate checking code
GetServerID was actually never meant to be for UI, it was a bug that has been fixed in the SDK. GetServerID is therefore no longer available. We should in the future however look for a solution to query this from the server, or have it actually stored on the client somewhere like the actual retail server id's.