Rework Halo_GotoNextPlaylist
Function CreateServer() has been fully removed from server context, it was originally only meant to be used by UI but it had been shared with the server for unknown reasons. In the server we should always stick to GameRules_ChangeMap(). Also, the server name and description is now always synced with the convar 'hostname' and 'hostdesc', so this should now work properly with the correct server name and server description.