Refactor LoadLobbyAfterLeave
We can just run "LeaveMatch" on listen server builds too, and then call CreateServer(). This seems to fix the rare crashing that occurs when this is ran as previously we also ran "disconnect" which brings us to the title screen. But going to the title screen while also changing levels could cause crashes in the snapshot worker. The same happens when DestroyServer() is called (on listen server).

Just send LeaveMatch to notify the server so script code could deal with it, and changelevel to mp_lobby with CreateServer().