Rework IsPlayerEliminated
Fix a bug where if player was eliminated, then disconnected, it was allowed to spectate any team. Force him to keep spectating the team if this one is still alive.

If player disconnects in Playing game state, mark him as Eliminated, to save his team and avoid him to spectate other teams if reconnected.

Only set TEAM_SPECTATOR if player connected later in the match but the player wasn't eliminated