fix B1D2D6 crash in 1v1 mode
Group removal, opponent being send to the lobby and group's ring destruction is already handled in the main thread, let it to be managed from main thread. The crash was caused for trying to remove the group twice in the same frame.

It makes sense why I decided to not run this part in scenarios mode, when this is the only function those modes share.

Warning: 1v1 mode could not be B1D2D6 free, as there could be more cases that trigger it, Challenges system needs to be audited if it keeps happening.